+++ perception/util/converter/detected_object_feature_remover/launch/detected_object_feature_remover.xml
@@ -2,7 +2,7 @@
   <arg name="input" />
   <arg name="output" />
 
-  <node pkg="dynamic_object_converter" exec="dynamic_object_converter" name="dynamic_object_converter" output="screen">
+  <node pkg="detected_object_feature_remover" exec="detected_object_feature_remover" name="detected_object_feature_remover" output="screen">
     <remap from="~/input" to="$(var input)"/>
     <remap from="~/output" to="$(var output)"/>
   </node>
